Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -82,7 +82,7 @@ def get_average_age_for_season(data, season)
           new_array.push(age)
 
       }
-     average_age = new_array.inject{ |sum, el| sum + el }.to_i / new_array.length
+     average_age = new_array.inject{ |sum, el| sum + el }.to_f / new_array.length
      return average_age
   end
   }
